refactor(SettingsPage): move logic with app language switching into new AppLanguageSwitchWidget
